refactor(economy): phase 3 slice 4 core scenario migration (#3939)
Migrate seven remaining imperative core suites (39 rows) to table-driven
scenario modules in test_support; economy test/ LOC 3,093 → 2,415.

Refs #3939
