+++ packages/colonizethis_economy/test/cost_check_test.dart
@@ -1,67 +1,9 @@
-import 'package:colonizethis_economy/colonizethis_economy.dart';
-import 'package:colonizethis_test/test.dart';
-
 import 'package:colonizethis_economy_test_support/colonizethis_economy_test_support.dart';
 
 void main() {
-  group('checkPreconditionsInOrder (Refs #3517 Cluster 2)', () {
-    test('returns null when every check passes', () {
-      final reason = checkPreconditionsInOrder([
-        (failReason: 'a', check: () => true),
-        (failReason: 'b', check: () => true),
-        (failReason: 'c', check: () => true),
-      ]);
-      expect(reason, isNull);
-    });
-
-    test('returns the first failing reason in list order', () {
-      final reason = checkPreconditionsInOrder([
-        (failReason: 'tech', check: () => true),
-        (failReason: 'workers', check: () => false),
-        (failReason: 'treasury', check: () => false),
-      ]);
-      expect(reason, 'workers');
-    });
-
-    test('honours canonical priority: earlier failure wins over later', () {
-      final reason = checkPreconditionsInOrder([
-        (failReason: 'tech', check: () => false),
-        (failReason: 'materials', check: () => false),
-      ]);
-      expect(reason, 'tech');
-    });
-
-    test('short-circuits: no later check runs once one fails', () {
-      final evaluated = <String>[];
-      final reason = checkPreconditionsInOrder([
-        (
-          failReason: 'first',
-          check: () {
-            evaluated.add('first');
-            return true;
-          },
-        ),
-        (
-          failReason: 'second',
-          check: () {
-            evaluated.add('second');
-            return false;
-          },
-        ),
-        (
-          failReason: 'third',
-          check: () {
-            evaluated.add('third');
-            return true;
-          },
-        ),
-      ]);
-      expect(reason, 'second');
-      expect(evaluated, ['first', 'second']);
-    });
-
-    test('empty precondition list passes (returns null)', () {
-      expect(checkPreconditionsInOrder(const []), isNull);
-    });
-  });
+  runLabeledScenarioGroup(
+    'checkPreconditionsInOrder (Refs #3517 Cluster 2)',
+    checkPreconditionsInOrderScenarios(),
+    runCheckPreconditionsInOrderScenario,
+  );
 }
