+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -15,7 +15,9 @@ export 'src/boycott_blocked_commodities_test_support.dart';
 export 'src/carry_forward_bid_notional_scenarios.dart';
 export 'src/consumption_phases_scenarios.dart';
 export 'src/consumption_scenarios.dart';
+export 'src/cost_check_scenarios.dart';
 export 'src/core_economy_test_support.dart';
+export 'src/economy_extraction_scenarios.dart';
 export 'src/economy_production_scenarios.dart';
 export 'src/deal_matcher_frr_scenarios.dart';
 export 'src/deal_matcher_priority_scenarios.dart';
@@ -28,6 +30,7 @@ export 'src/frr_credits_scenarios.dart';
 export 'src/frr_credits_test_support.dart';
 export 'src/frr_issue_ac_d5_scenarios.dart';
 export 'src/first_right_profit_scenarios.dart';
+export 'src/game_lookup_helpers_scenarios.dart';
 export 'src/gp_treasury_credit_accumulator_scenarios.dart';
 export 'src/lock_recovery_minor_bids_scenarios.dart';
 export 'src/lock_recovery_minor_bids_test_support.dart';
@@ -38,13 +41,17 @@ export 'src/non_gp_extraction_scenarios.dart';
 export 'src/non_gp_extraction_test_support.dart';
 export 'src/resource_extractor_scenarios.dart';
 export 'src/scenario_runner.dart';
+export 'src/sea_transport_scenarios.dart';
 export 'src/sellable_quantity_scenarios.dart';
 export 'src/resource_extractor_test_support.dart';
 export 'src/staged_bid_spend_scenarios.dart';
+export 'src/tile_extraction_pipeline_scenarios.dart';
 export 'src/tile_map_test_support.dart';
+export 'src/trade_cargo_capacity_scenarios.dart';
 export 'src/treasury_available_scenarios.dart';
 export 'src/player_context_scenarios.dart';
 export 'src/price_discovery_scenarios.dart';
+export 'src/projected_cost_engine_scenarios.dart';
 export 'src/purchased_tile_index_scenarios.dart';
 export 'src/purchased_tile_index_test_support.dart';
 export 'src/purchased_tile_riches_scenarios.dart';
